🚚 Rename: side-menu로 수정
